Move src/src ... powermodule to new structure
